refactor(ui): rename separator and toggle components for consistency
