+++ src/components/MenuBarSeparator/MenuBarSeparator.tsx
